requirements.txt
Pillow
Click
pillow-heif